agent-gateway/pyproject.toml
[project]
name = "agent-gateway"
version = "0.1"
description = "AIGIS Governance Agent Gateway"
authors = [
    { name = "Pedro Leta", email = "pedrogleta@gmail.com" }
]
license = { text = "Apache License 2.0" }
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=1.5.0",
    "immutabledict>=4.2.1",
    "sqlglot>=26.10.1",
    "db-dtypes>=1.4.2",
    "regex>=2024.11.6",
    "tabulate>=0.9.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "absl-py>=2.2.2",
    "pydantic>=2.11.3",
    "pandas>=2.3.0",
    "numpy>=2.3.1",
    "pytest>=8.4.1",
    "google-cloud-bigquery>=3.34.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "opik>=1.8.23",
    "llm-sandbox[docker]>=0.3.17",
    "minio>=7.2.0",
    "uuid>=1.30",
]
requires-python = ">=3.12"

